+++ lib/src/extensions/extensions.dart
@@ -1,8 +1,3 @@
-export 'T_watcher_extension.dart';
-export 'date_time_watcher_extension.dart';
-export 'list_watcher_extension.dart';
 export 'listenable_extensions.dart';
-export 'map_watcher_extension.dart';
 export 'stream_to_watcher.dart';
-export 'uri_watcher_extension.dart';
 export 'watcher_extensions.dart';